code/requirements.txt
aiosqlite==0.21.0
beautifulsoup4==4.13.4
chainlit==2.4.0
chromadb==1.0.12
fastapi==0.115.9
feedparser==6.0.11
langchain==0.3.25
langchain_chroma==0.2.4
langchain_community==0.3.25
langchain_core==0.3.65
langchain_openai==0.3.23
markdown==3.8.2
plotly==5.20.0
Requests==2.32.4
tqdm==4.66.5
unstructured==0.17.2